Front: add option to customize customer information forms
Add option to change form groups for Customer Information, Company Information and Address Book forms

This enables users to customize specially the fields which are needed for specific solutions and hide forms when not
needed

Refs MOOS-46
