+++ shuup/front/apps/customer_information/settings.py
@@ -1,0 +1,15 @@
+# This file is part of Shuup.
+#
+# Copyright (c) 2012-2019, Shoop Commerce Ltd. All rights reserved.
+#
+# This source code is licensed under the OSL-3.0 license found in the
+# LICENSE file in the root directory of this source tree.
+
+#: The default customer information edit form
+SHUUP_CUSTOMER_INFORMATION_EDIT_FORM = "shuup.front.apps.customer_information.forms:CustomerInformationFormGroup"
+
+#: The default company information edit form
+SHUUP_COMPANY_INFORMATION_EDIT_FORM = "shuup.front.apps.customer_information.forms:CompanyInformationFormGroup"
+
+#: The default address book edit form
+SHUUP_ADDRESS_BOOK_EDIT_FORM = "shuup.front.apps.customer_information.forms:AddressBookFormGroup"
